scripts/training_requirements.txt
accelerate==1.6.0
bitsandbytes==0.45.4
scipy==1.15.2
sentencepiece==0.2.0
einops==0.8.1
wandb==0.19.10
deepspeed==0.15.4
typer==0.9.0
tensorboard==2.19.0
transformers==4.51.3
liger-kernel==0.5.9
peft==0.15.1
datasets==3.5.0
trl==0.18.0
docker
toml
debugpy
textstat==0.7.8
langcheck==0.9.0
detoxify
mlflow
aiohttp
requests
toml
tenacity
minio
PyYAML
pandas==2.2.3
mpi4py